+++ apps/extension/src/hooks/onboarding.ts
@@ -3,7 +3,7 @@ import { passwordSelector } from '../state/password';
 import { generateSelector } from '../state/seed-phrase/generate';
 import { importSelector } from '../state/seed-phrase/import';
 import { walletsSelector } from '../state/wallets';
-import { swClient } from '../routes/service-worker/internal/client/internal';
+import { swClient } from '../routes/service-worker/internal/client';
 
 // Saves hashed password, uses that hash to encrypt the seed phrase
 // and then saves that to session + local storage
