+++ apps/extension/src/popup.tsx
@@ -2,10 +2,11 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { popupRouter } from './routes/popup/router';
 import { RouterProvider } from 'react-router-dom';
-import { QueryClientProvider } from '@tanstack/react-query';
-import { queryClient } from './services/clients';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 import 'ui/styles/globals.css';
+
+export const queryClient = new QueryClient();
 
 const startPopup = () => {
   const rootElement = document.getElementById('popup-root') as HTMLDivElement;
