+++ apps/extension/src/routes/service-worker/internal/routes/clear-cache.ts
@@ -1,5 +1,5 @@
-import { services } from '../../../service-worker';
-import { SwMessage, SwMessageHandler } from './types';
+import { services } from '../../../../service-worker';
+import { SwMessage, SwMessageHandler } from '../types';
 
 export type ClearCacheMessage = SwMessage<'CLEAR_CACHE', undefined, Promise<void>>;
