+++ apps/extension/src/routes/service-worker/root-router.ts
@@ -1,7 +1,6 @@
 import { ServiceWorkerResponse } from './internal/types';
 import { internalRouter, isInternalRequest, SwRequestMessage } from './internal/router';
-import { isViewServerReq } from './view-protocol-server/helpers/generic';
-import { viewServerRouter } from './view-protocol-server/router';
+import { isViewServerReq, viewServerRouter } from './view-protocol-server/router';
 
 // Used to filter for service worker messages and narrow their type to pass to the typed handler.
 // Exposed to service worker for listening for internal and external messages
@@ -15,7 +14,7 @@ export const swMessageHandler = (
   if (isInternalRequest(message)) {
     return internalRouter(message, sendResponse);
   } else if (isViewServerReq(message)) {
-    viewServerRouter(message, sender, sendResponse);
+    viewServerRouter(message, sender);
   }
   return;
 };
