+++ apps/extension/src/routes/service-worker/view-protocol-server/balances.ts
@@ -4,11 +4,11 @@ import {
   SpendableNoteRecord,
 } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/view/v1alpha1/view_pb';
 import { addLoHi, Base64Str, uint8ArrayToBase64 } from 'penumbra-types';
-import { ViewReqMessage } from './helpers/generic';
 import { AddressIndex } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/keys/v1alpha1/keys_pb';
 import { Value } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1alpha1/asset_pb';
 import { Amount } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/num/v1alpha1/num_pb';
 import { services } from '../../../service-worker';
+import { ViewReqMessage } from './router';
 
 type AssetIdStr = Base64Str;
 type BalancesMap = Record<AssetIdStr, BalancesResponse>;
