+++ apps/extension/src/routes/service-worker/view-protocol-server/tx-info-by-hash.ts
@@ -1,10 +1,10 @@
-import { ViewReqMessage } from './helpers/generic';
 import {
   TransactionInfoByHashRequest,
   TransactionInfoByHashResponse,
 } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/view/v1alpha1/view_pb';
 import { services } from '../../../service-worker';
 import { NoteSource } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/component/chain/v1alpha1/chain_pb';
+import { ViewReqMessage } from './router';
 
 export const isTxInfoByHashRequest = (req: ViewReqMessage): req is TransactionInfoByHashRequest => {
   return req.getType().typeName === TransactionInfoByHashRequest.typeName;
