+++ apps/extension/src/routes/service-worker/view-protocol-server/tx-info.ts
@@ -3,9 +3,9 @@ import {
   TransactionInfoRequest,
   TransactionInfoResponse,
 } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/view/v1alpha1/view_pb';
-import { ViewReqMessage } from './helpers/generic';
 
 import { services } from '../../../service-worker';
+import { ViewReqMessage } from './router';
 
 export const isTransactionInfoRequest = (msg: ViewReqMessage): msg is TransactionInfoRequest => {
   return msg.getType().typeName === TransactionInfoRequest.typeName;
