+++ apps/extension/src/routes/service-worker/view-protocol-server/wallet-id.ts
@@ -1,4 +1,3 @@
-import { ViewReqMessage } from './helpers/generic';
 import {
   WalletIdRequest,
   WalletIdResponse,
@@ -6,6 +5,7 @@ import {
 import { localExtStorage } from 'penumbra-storage';
 import { WalletId } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/keys/v1alpha1/keys_pb';
 import { stringToUint8Array } from 'penumbra-types';
+import { ViewReqMessage } from './router';
 
 export const isWalletIdRequest = (req: ViewReqMessage): req is WalletIdRequest => {
   return req.getType().typeName === WalletIdRequest.typeName;
